ORO-178 Update numPoints parameter type to size_t in CDF setup
- Change numPoints from int32_t to size_t for better type safety
- Adjust related calculations to use size_t for consistency

+++ src/galaxy/galaxy_cdf.h
@@ -77,7 +77,7 @@ class CumulativeDistributionFunction
                         float64_t bulgeRadius,
                         float64_t min,
                         float64_t max,
-                        int32_t   numPoints);
+                        size_t    numPoints);
 
     /**
      * @brief Get radius value from probability
